feat: move js files to bridge dir
