+++ libs/coin-polkadot/src/bridge/getFeesForTransaction.ts
@@ -1,10 +1,10 @@
 import { BigNumber } from "bignumber.js";
 import { getAbandonSeedAddress } from "@ledgerhq/cryptoassets";
-import type { PolkadotAccount, Transaction } from "./types";
-import { calculateAmount } from "./logic";
-import { buildTransaction } from "./js-buildTransaction";
-import { fakeSignExtrinsic } from "./js-signOperation";
-import { PolkadotAPI } from "./api";
+import type { PolkadotAccount, Transaction } from "../types";
+import { calculateAmount } from "../logic";
+import { buildTransaction } from "./buildTransaction";
+import { fakeSignExtrinsic } from "./signOperation";
+import { PolkadotAPI } from "../api";
 
 /**
  * Fetch the transaction fees for a transaction
