+++ libs/coin-polkadot/src/index.ts
@@ -1,1 +1,3 @@
 export * from "./types";
+
+export { buildCurrencyBridge, buildAccountBridge, createBridges } from "./bridge/index";
